cukinia-tests: add kernel minimization test
Add a test to check that the kernel size is less than a
specified maximum size. This is useful to ensure that the kernel
is optimized and does not contain unnecessary features.
This test is checking for kernel images named `vmlinuz-<version>`
or `Image` in the `/boot` directory. The other image names are not
considered as they are compressed and not relevant for this test.
In this case, the test is skipped

SFL: #17874
Change-Id: I3b467c13fac115bce62077e8b5c104c06271be9c
Signed-off-by: Kévin L'hôpital <kevin.lhopital@savoirfairelinux.com>